Chapter 7/requirements.txt
langchain==0.1.11
langchain-experimental==0.0.53
langchain-openai==0.0.6
python-dotenv==1.0.0
chromadb==0.4.22
pypdf==3.17.3
openai==1.12.0
tiktoken==0.5.2
streamlit==1.32.2
docarray==0.40.0
pandas==2.2.1